fix/safe_app: integrate Java bindings with safe_client_libs master
- Also update variables and parameters naming for consistency.

+++ ffi_utils/src/bindgen_utils.rs
@@ -35,19 +35,17 @@ pub fn copy_files<S: AsRef<Path>, T: AsRef<Path>>(
     for entry in WalkDir::new(source) {
         let entry = entry?;
 
-        if entry.path().is_file() {
-            if entry
+        if entry.path().is_file() &&
+            entry
                 .path()
                 .to_str()
                 .map(|s| s.ends_with(extension))
                 .unwrap_or(false)
-            {
-                let source_path = entry.path();
-                let target_path =
-                    target.join(source_path.strip_prefix(source).unwrap_or(&source_path));
+        {
+            let source_path = entry.path();
+            let target_path = target.join(source_path.strip_prefix(source).unwrap_or(source_path));
 
-                let _ = fs::copy(source_path, target_path)?;
-            }
+            let _ = fs::copy(source_path, target_path)?;
         }
     }
